+++ api/pkg/resources/resources.go
@@ -1,9 +1,120 @@
 package resources
 
 import (
+	"fmt"
+
+	"github.com/golang/glog"
+	apiv1 "github.com/kubermatic/kubermatic/api/pkg/api/v1"
+	kubermaticv1 "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
+	"github.com/kubermatic/kubermatic/api/pkg/provider"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	corev1lister "k8s.io/client-go/listers/core/v1"
+)
+
+const (
+	//AddonManagerDeploymentName is the name for the addon-manager deployment
+	AddonManagerDeploymentName = "addon-manager"
+	//ApiserverDeploymenName is the name for the apiserver deployment
+	ApiserverDeploymenName = "apiserver"
+	//ControllerManagerDeploymentName is the name for the controller manager deployment
+	ControllerManagerDeploymentName = "controller-manager"
+	//SchedulerDeploymentName is the name for the scheduler deployment
+	SchedulerDeploymentName = "scheduler"
+	//MachineControllerDeploymentName is the name for the machine-controller deployment
+	MachineControllerDeploymentName = "machine-controller"
+	//OpenVPNServerDeploymentName is the name for the openvpn server deployment
+	OpenVPNServerDeploymentName = "openvpn-server"
+
+	//PrometheusStatefulSetName is the name for the prometheus StatefulSet
+	PrometheusStatefulSetName = "prometheus"
+	//EtcdStatefulSetName is the name for the etcd StatefulSet
+	EtcdStatefulSetName = "etcd"
+
+	//ApiserverExternalServiceName is the name for the external apiserver service
+	ApiserverExternalServiceName = "apiserver-external"
+	//ApiserverInternalServiceName is the name for the internal apiserver service
+	ApiserverInternalServiceName = "apiserver"
+	//PrometheusServiceName is the name for the prometheus service
+	PrometheusServiceName = "prometheus"
+	//EtcdServiceName is the name for the etcd service
+	EtcdServiceName = "etcd"
+	//EtcdClientServiceName is the name for the etcd service
+	EtcdClientServiceName = "etcd-client"
+	//OpenVPNServerServiceName is the name for the openvpn server service
+	OpenVPNServerServiceName = "openvpn-server"
+
+	//AdminKubeconfigSecretName is the name for the secret containing the private ca key
+	AdminKubeconfigSecretName = "admin-kubeconfig"
+	//CAKeySecretName is the name for the secret containing the private ca key
+	CAKeySecretName = "ca-key"
+	//CACertSecretName is the name for the secret containing the ca.crt
+	CACertSecretName = "ca-cert"
+	//ApiserverTLSSecretName is the name for the secrets required for the apiserver tls
+	ApiserverTLSSecretName = "apiserver-tls"
+	//KubeletClientCertificatesSecretName is the name for the secret containing the kubelet client certificates
+	KubeletClientCertificatesSecretName = "kubelet-client-certificates"
+	//ServiceAccountKeySecretName is the name for the secret containing the service account key
+	ServiceAccountKeySecretName = "service-account-key"
+	//TokensSecretName is the name for the secret containing the user tokens
+	TokensSecretName = "tokens"
+	//OpenVPNServerCertificatesSecretName is the name for the secret containing the openvpn server certificates
+	OpenVPNServerCertificatesSecretName = "openvpn-server-certificates"
+	//OpenVPNClientCertificatesSecretName is the name for the secret containing the openvpn client certificates
+	OpenVPNClientCertificatesSecretName = "openvpn-client-certificates"
+
+	//CloudConfigConfigMapName is the name for the configmap containing the cloud-config
+	CloudConfigConfigMapName = "cloud-config"
+	//OpenVPNClientConfigConfigMapName is the name for the configmap containing the openvpn client config used within the user cluster
+	OpenVPNClientConfigConfigMapName = "openvpn-client-configs"
+	//PrometheusConfigConfigMapName is the name for the configmap containing the prometheus config
+	PrometheusConfigConfigMapName = "prometheus"
+
+	//PrometheusServiceAccountName is the name for the Prometheus serviceaccount
+	PrometheusServiceAccountName = "prometheus"
+
+	//PrometheusRoleName is the name for the Prometheus role
+	PrometheusRoleName = "prometheus"
+
+	//PrometheusRoleBindingName is the name for the Prometheus rolebinding
+	PrometheusRoleBindingName = "prometheus"
+
+	// DefaultOwnerReadOnlyMode represents file mode 0400 in decimal
+	DefaultOwnerReadOnlyMode = 256
+
+	// AppLabelKey defines the label key app which should be used within resources
+	AppLabelKey = "app"
+)
+
+const (
+	// CAKeySecretKey ca.key
+	CAKeySecretKey = "ca.key"
+	// CACertSecretKey ca.crt
+	CACertSecretKey = "ca.crt"
+	// ApiserverTLSKeySecretKey apiserver-tls.key
+	ApiserverTLSKeySecretKey = "apiserver-tls.key"
+	// ApiserverTLSCertSecretKey apiserver-tls.crt
+	ApiserverTLSCertSecretKey = "apiserver-tls.crt"
+	// KubeletClientKeySecretKey kubelet-client.key
+	KubeletClientKeySecretKey = "kubelet-client.key"
+	// KubeletClientCertSecretKey kubelet-client.crt
+	KubeletClientCertSecretKey = "kubelet-client.crt"
+	// ServiceAccountKeySecretKey sa.key
+	ServiceAccountKeySecretKey = "sa.key"
+	// AdminKubeconfigSecretKey admin-kubeconfig
+	AdminKubeconfigSecretKey = "admin-kubeconfig"
+	// TokensSecretKey tokens.csv
+	TokensSecretKey = "tokens.csv"
+	// OpenVPNServerKeySecretKey server.key
+	OpenVPNServerKeySecretKey = "server.key"
+	// OpenVPNServerCertSecretKey server.crt
+	OpenVPNServerCertSecretKey = "server.crt"
+	// OpenVPNInternalClientKeySecretKey client.key
+	OpenVPNInternalClientKeySecretKey = "client.key"
+	// OpenVPNInternalClientCertSecretKey client.crt
+	OpenVPNInternalClientCertSecretKey = "client.crt"
 )
 
 // ConfigMapCreator defines an interface to create/update ConfigMap's
@@ -29,3 +140,114 @@ type ClusterRoleBindingCreator = func(data *TemplateData, existing *rbacv1.Clust
 
 // DeploymentCreator defines an interface to create/update Deployment's
 type DeploymentCreator = func(data *TemplateData, existing *appsv1.Deployment) (*appsv1.Deployment, error)
+
+// TemplateData is a group of data required for template generation
+type TemplateData struct {
+	Cluster           *kubermaticv1.Cluster
+	Version           *apiv1.MasterVersion
+	DC                *provider.DatacenterMeta
+	SecretLister      corev1lister.SecretLister
+	ConfigMapLister   corev1lister.ConfigMapLister
+	ServiceLister     corev1lister.ServiceLister
+	OverwriteRegistry string
+	NodePortRange     string
+}
+
+// GetClusterRef returns a instance of a OwnerReference for the Cluster in the TemplateData
+func (d *TemplateData) GetClusterRef() metav1.OwnerReference {
+	gv := kubermaticv1.SchemeGroupVersion
+	return *metav1.NewControllerRef(d.Cluster, gv.WithKind("Cluster"))
+}
+
+// Int32 returns a pointer to the int32 value passed in.
+func Int32(v int32) *int32 {
+	return &v
+}
+
+// Int64 returns a pointer to the int64 value passed in.
+func Int64(v int64) *int64 {
+	return &v
+}
+
+// Bool returns a pointer to the bool value passed in.
+func Bool(v bool) *bool {
+	return &v
+}
+
+// String returns a pointer to the string value passed in.
+func String(v string) *string {
+	return &v
+}
+
+// NewTemplateData returns an instance of TemplateData
+func NewTemplateData(
+	cluster *kubermaticv1.Cluster,
+	version *apiv1.MasterVersion,
+	dc *provider.DatacenterMeta,
+	secretLister corev1lister.SecretLister,
+	configMapLister corev1lister.ConfigMapLister,
+	serviceLister corev1lister.ServiceLister,
+	overwriteRegistry string,
+	nodePortRange string) *TemplateData {
+	return &TemplateData{
+		Cluster:           cluster,
+		DC:                dc,
+		Version:           version,
+		ConfigMapLister:   configMapLister,
+		SecretLister:      secretLister,
+		ServiceLister:     serviceLister,
+		OverwriteRegistry: overwriteRegistry,
+		NodePortRange:     nodePortRange,
+	}
+}
+
+// SecretRevision returns the resource version of the secret specified by name. A empty string will be returned in case of an error
+func (d *TemplateData) SecretRevision(name string) (string, error) {
+	secret, err := d.SecretLister.Secrets(d.Cluster.Status.NamespaceName).Get(name)
+	if err != nil {
+		return "", fmt.Errorf("could not get secret %s from lister for cluster %s: %v", name, d.Cluster.Name, err)
+	}
+	return secret.ResourceVersion, nil
+}
+
+// ConfigMapRevision returns the resource version of the configmap specified by name. A empty string will be returned in case of an error
+func (d *TemplateData) ConfigMapRevision(name string) (string, error) {
+	cm, err := d.ConfigMapLister.ConfigMaps(d.Cluster.Status.NamespaceName).Get(name)
+	if err != nil {
+		return "", fmt.Errorf("could not get configmap %s from lister for cluster %s: %v", name, d.Cluster.Name, err)
+	}
+	return cm.ResourceVersion, nil
+}
+
+// ProviderName returns the name of the clusters providerName
+func (d *TemplateData) ProviderName() string {
+	p, err := provider.ClusterCloudProviderName(d.Cluster.Spec.Cloud)
+	if err != nil {
+		glog.V(0).Infof("could not identify cloud provider: %v", err)
+	}
+	return p
+}
+
+// GetApiserverExternalNodePort returns the nodeport of the external apiserver service
+func (d *TemplateData) GetApiserverExternalNodePort() (int32, error) {
+	s, err := d.ServiceLister.Services(d.Cluster.Status.NamespaceName).Get(ApiserverExternalServiceName)
+	if err != nil {
+
+		return 0, fmt.Errorf("failed to get NodePort for external apiserver service: %v", err)
+
+	}
+	return s.Spec.Ports[0].NodePort, nil
+}
+
+// ImageRegistry returns the image registry to use or the passed in default if no override is specified
+func (d *TemplateData) ImageRegistry(defaultRegistry string) string {
+	if d.OverwriteRegistry != "" {
+		return d.OverwriteRegistry
+	}
+	return defaultRegistry
+}
+
+// GetLabels returns default labels every resource should have
+func GetLabels(app string) map[string]string {
+	return map[string]string{AppLabelKey: app}
+}
